chore(codegen): cleanup
