+++ internal/entity/openapi.spec.yaml
