+++ internal/entity/teams_test.go
@@ -15,49 +15,38 @@
 // CONSEQUENTIAL, SPECIAL, INCIDENTAL, INDIRECT, OR DIRECT DAMAGES, HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
 // ARISING OUT OF THIS AGREEMENT. THE FOREGOING SHALL APPLY TO THE EXTENT PERMITTED BY APPLICABLE LAW.
 
-package entities_test
+package entity_test
 
 import (
 	"testing"
 
-	"go.breu.io/ctrlplane/internal/entities"
+	"github.com/gosimple/slug"
+
+	"go.breu.io/ctrlplane/internal/entity"
 	"go.breu.io/ctrlplane/internal/shared"
 )
 
-var (
-	password string
-)
-
-func TestUser(t *testing.T) {
-	password = "password"
-	user := &entities.User{Password: password}
-	_ = user.PreCreate()
+func TestTeam(t *testing.T) {
+	team := &entity.Team{
+		Name: "Team Name",
+	}
+	_ = team.PreCreate()
 
 	opsTests := shared.TestFnMap{
-		"SetPassword":    shared.TestFn{Args: user, Want: nil, Run: testUserSetPassword},
-		"VerifyPassword": shared.TestFn{Args: user, Want: nil, Run: testUserVerifyPassword},
+		"Slug": shared.TestFn{Args: team, Want: nil, Run: testTeamSlug},
 	}
 
-	t.Run("GetTable", testEntityGetTable("users", user))
-	t.Run("EntityOps", testEntityOps(user, opsTests))
+	t.Run("GetTable", testEntityGetTable("teams", team))
+	t.Run("EntityOps", testEntityOps(team, opsTests))
 }
 
-func testUserSetPassword(args interface{}, want interface{}) func(*testing.T) {
-	user := args.(*entities.User)
+func testTeamSlug(args interface{}, want interface{}) func(*testing.T) {
+	team := args.(*entity.Team)
+	sluglen := len(slug.Make(team.Name)) + 1 + 22
 
 	return func(t *testing.T) {
-		if user.Password == "password" {
-			t.Errorf("expected password to be encrypted")
-		}
-	}
-}
-
-func testUserVerifyPassword(args interface{}, want interface{}) func(*testing.T) {
-	v := args.(*entities.User)
-
-	return func(t *testing.T) {
-		if !v.VerifyPassword(password) {
-			t.Errorf("expected password to be verified")
+		if len(team.Slug) != sluglen {
+			t.Errorf("slug length is not correct, got: %d, want: %d", len(team.Slug), sluglen)
 		}
 	}
 }
