+++ internal/entity/users_test.go
@@ -15,15 +15,49 @@
 // CONSEQUENTIAL, SPECIAL, INCIDENTAL, INDIRECT, OR DIRECT DAMAGES, HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
 // ARISING OUT OF THIS AGREEMENT. THE FOREGOING SHALL APPLY TO THE EXTENT PERMITTED BY APPLICABLE LAW.
 
-package entities_test
+package entity_test
 
 import (
 	"testing"
 
-	"go.breu.io/ctrlplane/internal/entities"
+	"go.breu.io/ctrlplane/internal/entity"
+	"go.breu.io/ctrlplane/internal/shared"
 )
 
-func TestTeamUser(t *testing.T) {
-	teamUser := &entities.TeamUser{}
-	t.Run("GetTable", testEntityGetTable("team_users", teamUser))
+var (
+	password string
+)
+
+func TestUser(t *testing.T) {
+	password = "password"
+	user := &entity.User{Password: password}
+	_ = user.PreCreate()
+
+	opsTests := shared.TestFnMap{
+		"SetPassword":    shared.TestFn{Args: user, Want: nil, Run: testUserSetPassword},
+		"VerifyPassword": shared.TestFn{Args: user, Want: nil, Run: testUserVerifyPassword},
+	}
+
+	t.Run("GetTable", testEntityGetTable("users", user))
+	t.Run("EntityOps", testEntityOps(user, opsTests))
+}
+
+func testUserSetPassword(args interface{}, want interface{}) func(*testing.T) {
+	user := args.(*entity.User)
+
+	return func(t *testing.T) {
+		if user.Password == "password" {
+			t.Errorf("expected password to be encrypted")
+		}
+	}
+}
+
+func testUserVerifyPassword(args interface{}, want interface{}) func(*testing.T) {
+	v := args.(*entity.User)
+
+	return func(t *testing.T) {
+		if !v.VerifyPassword(password) {
+			t.Errorf("expected password to be verified")
+		}
+	}
 }
