+++ internal/providers/github/openapi.gen.go
@@ -156,6 +156,7 @@ type ServerInterfaceWrapper struct {
 
 func (w *ServerInterfaceWrapper) GithubCompleteInstallation(ctx echo.Context) error {
 	var err error
+
 	ctx.Set(BearerAuthScopes, []string{""})
 
 	ctx.Set(APIKeyAuthScopes, []string{""})
@@ -172,6 +173,7 @@ func (w *ServerInterfaceWrapper) GithubCompleteInstallation(ctx echo.Context) er
 
 func (w *ServerInterfaceWrapper) GithubGetRepos(ctx echo.Context) error {
 	var err error
+
 	ctx.Set(BearerAuthScopes, []string{""})
 
 	ctx.Set(APIKeyAuthScopes, []string{""})
@@ -196,9 +198,8 @@ func (w *ServerInterfaceWrapper) GithubWebhook(ctx echo.Context) error {
 	return err
 }
 
-// This is a simple interface which specifies echo.Route addition functions which
-// are present on both echo.Echo and echo.Group, since we want to allow using
-// either of them for path registration
+// EchoRouter is an interface that wraps the methods of echo.Echo & echo.Group to provide a common interface
+// for registering routes.
 type EchoRouter interface {
 	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
 	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
